Use yarnpkg to lock our dependencies
This change switch from npm to [yarnpkg](https://yarnpkg.com/).

An extension is a kind of self-contained application unlike a library.
So I think it's reasonable to add the lock file into the repository
for dependency management.

Alternative Approach
--------------------

__Use npm v5 or later__

This plan don't requires to install yarnpkg for our contributor.
However, it specifies npm version. npm is bundled with nodejs.
I'm not sure to that it's safe to install some specific version of npm
into user local machine.

And I also think it's a little hard to specify the version of npm
because nodejs also update bundled npm on upgrading nodejs.

If we can use nodejs 8 or later, this plan might be reasonable because
nodejs 8 bundles npm v5.x. But we use nodejs 7. Thus I propose
to use the latest yarnpkg _at this time_.