Don't add default build tasks to task.json on opening a workspace
Motivation
----------------

The previous behavior always create `.vscode/tasks.json` if there is no it.

If the workspace does not contains `.vscode/tasks.json` intentionally by
some reasons (it is not a root dir of the repository, the repository
has not ignored `.vscode` yet, or others), then the previous behavior
force to users to ignore/remove `.vscode/tasks.json` on creating their changesets for the repository.
I'd like to stop it.