Prevent error when rustc is not in $PATH
...by extending `$PATH` with a common place where rustup drops binaries.

Story time! I keep having this weird issue when opening a Rust project
in vscode:

```
TypeError: Cannot read property 'toString' of null
	at makeRlsEnv (/Users/pascal/.vscode-insiders/extensions/rust-lang.rust-0.2.2/out/src/extension.js:44:36)
	at makeRlsProcess (/Users/pascal/.vscode-insiders/extensions/rust-lang.rust-0.2.2/out/src/extension.js:55:17)
	at autoUpdate.then (/Users/pascal/.vscode-insiders/extensions/rust-lang.rust-0.2.2/out/src/extension.js:114:57)
	at <anonymous>
```

Debugging the extension lead me to `src/extension.ts:41` [1], which
assumes that the return value of `spawnSync` has a `stdout` property
that can be converted to a string. Sadly, this is not always the case:
When `spawn` could not find the binary to execute, it returns an object
with an `error` property and also `stdout: null`.

I have no clue why it can't find `rustc` on my machine, but I sure have
a weird setup. I'm using macOS, my default shell is fish, and this is in
version `1.16.0-insider` of vscode. I'm using rustup, and all
rust-related binaries are only available locally in `~/.cargo/bin/` from
what I can tell.

This change should not cause any trouble---but I have only tested it on
macOS. I'm not sure how other system (Windows) handles `$PATH` or even
what syntax it expects. If vscode-rust targets Windows, patches to this
may be required.

[1]: https://github.com/rust-lang-nursery/rls-vscode/blob/63000e65870f18318615c27a72c683cec05a3cbf/src/extension.ts#L41