Add "on-enter" rules for comments
This adds the convenience of VSCode's hooks for onEnter events for Rust.
Now, when pressing `enter` within a  comment, the comment continues. For
example: after typing `///` and pressing `enter`, the next line that the
cursor is on will also have `///`, followed by a single space.

`//!`, `/**`, and `/*!` multi-line comments are also supported. Single
line comments (`//`), are not continued on `enter`.