Allow having Cargo.toml in a subdirectory with `rust-client.cargoSubdir`
Fixes https://github.com/rust-lang-nursery/rls-vscode/issues/222.

This patch is quite hacky and may reveal my lack of knowledge on the RLS or
LSP architectures. My idea was to give the user the opportunity using the
variable `rust-client.cargoSubdir` to select a subdirectory as the root for
the Cargo.toml file.

To do that, I hacked the vscode's workspaceFoldervariable; my hack won't
work with multi-workspaces.

Alternate solutions:
- (not so good idea) add a new setting to the initialize JSON-RPC (similar
  to the existing `omitInitBuild` in the rls server) that would set a
  different path that the one given by vscode's workspaceFolder.
- (better) we add support for multi-root workspaces in the RLS server;
  (mentionned in  <https://github.com/rust-lang-nursery/rls/issues/608>)
  but this does not really help when we want to open a general project
  where one of the subfolders is a rust project.
- we tell people that in order to get RLS working in a subdirectory, they
  must use the 'Multi-crate projects' feature or the 'rlsCommandOverride'
  feature (see README at <https://github.com/mehcode/atom-ide-rust>).