Allow to not look for config in outermost nested folder.
This patch makes it configurable where the extension looks for Rust
config relevant to a file in case of nested top-level directories.

Example: imagine a multi-language project in `<project_path>` with a
Rust component under `<project_path>/<rust subdir>`. In this case a
developer might open a workspace that contains both `<project_path>` and
`<project_path>/<rust subdir>` as two different top-level directories
(the former for easy access to the Rust code they work on, the latter to
occasionally look up interfaces and docs from outside of the Rust
directory.

This setup doesn't work without this patch: the extension is looking for
Rust config always in the outermost of the possibly multiple workspace
folders that contain the file.

This patch makes the behavior configurable, so that for the setup above
one can set:

  `"rust-client.nestedMultiRootConfigInOutermost": false`

and have the extension working in `<rust subdir>` despite also having
`<project path>` open.

Fixes #495.