Don't start servers for every already opened files
This models better what built-in TS/JS services do, in an attempt to be
more lazy and less resource-intensive.
This addresses a use case where the user may prefer to operate on a
monorepo with lots of opened files but keep the sessions short-ish,
where they don't touch every project in a single session.